Move io_lifetimes and libc modules out of imp::net and into imp.
