+++ src/imp/libc/net/mod.rs
@@ -1,11 +1,5 @@
 mod addr;
 mod ext;
-#[cfg(windows)]
-pub(crate) mod io_lifetimes;
-#[cfg(windows)]
-pub(crate) mod libc;
-#[cfg(not(windows))]
-pub(crate) use libc;
 mod read_sockaddr;
 mod send_recv;
 mod types;
