+++ src/imp/libc/net/write_sockaddr.rs
@@ -4,7 +4,7 @@
 
 use super::ext::{Ipv4AddrExt, Ipv6AddrExt, SocketAddrV6Ext};
 #[cfg(windows)]
-use super::libc;
+use super::super::libc;
 #[cfg(not(windows))]
 use super::SocketAddrUnix;
 use super::{SocketAddr, SocketAddrStorage};
